M8 Continued work on the Sales Reports
The visuals are in place, but a number of back-end things still need to
be taken care of.   I need to get the list to show up correctly (trying
to use code from the interest list, but it isn't working).  Once that is
done, I need to make the checks for only the right sales to show up on
the home screen.  Additional information for the sales reports is also
something that should be added (will need to mess with the product
class...).  Once this is all done, persistence needs to be added in.